adjust structure
