+++ src/bin/profiling.rs
@@ -11,40 +11,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-
-use std::{ptr, slice};
-
-use libc::{self, c_void, c_char};
-
-extern "C" {
-    #[cfg_attr(target_os = "macos", link_name = "je_malloc_stats_print")]
-    fn malloc_stats_print(write_cb: extern "C" fn(*mut c_void, *const c_char), cbopaque: *mut c_void, opts: *const c_char);
-}
-
-extern "C" fn write_cb(printer: *mut c_void, msg: *const c_char) {
-    unsafe {
-        let buf = &mut *{printer as *mut Vec<u8>};
-        let len = libc::strlen(msg);
-        let bytes = slice::from_raw_parts(msg as *const u8, len);
-        buf.extend_from_slice(bytes);
-    }
-}
-
-pub fn print_prof() {
-    let mut buf = Vec::new();
-    unsafe { malloc_stats_print(write_cb, &mut buf as *mut Vec<u8> as *mut c_void, ptr::null()) }
-    info!("{}", String::from_utf8_lossy(&buf));
-}
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn test_stats_print() {
-        // just print the data, ensure it doesn't core.
-        super::print_prof()
-    }
-}
-
 #[cfg(feature = "mem-profiling")]
 mod imp {
     use std::ffi::CString;
