+++ src/bin/signal_handler.rs
@@ -14,8 +14,10 @@
 
 #[cfg(unix)]
 mod imp {
+    use std::{ptr, slice};
     use std::sync::Arc;
 
+    use libc::{self, c_void, c_char};
     use rocksdb::DB;
 
     use tikv::server::Msg;
@@ -26,6 +28,26 @@ mod imp {
 
     const ROCKSDB_DB_STATS_KEY: &'static str = "rocksdb.dbstats";
     const ROCKSDB_CF_STATS_KEY: &'static str = "rocksdb.cfstats";
+
+    extern "C" {
+        #[cfg_attr(target_os = "macos", link_name = "je_malloc_stats_print")]
+        fn malloc_stats_print(write_cb: extern "C" fn(*mut c_void, *const c_char), cbopaque: *mut c_void, opts: *const c_char);
+    }
+
+    extern "C" fn write_cb(printer: *mut c_void, msg: *const c_char) {
+        unsafe {
+            let buf = &mut *{printer as *mut Vec<u8>};
+            let len = libc::strlen(msg);
+            let bytes = slice::from_raw_parts(msg as *const u8, len);
+            buf.extend_from_slice(bytes);
+        }
+    }
+
+    fn print_malloc_stats() {
+        let mut buf = Vec::new();
+        unsafe { malloc_stats_print(write_cb, &mut buf as *mut Vec<u8> as *mut c_void, ptr::null()) }
+        info!("{}", String::from_utf8_lossy(&buf));
+    }
 
     // TODO: remove backup_path from configuration
     pub fn handle_signal(ch: SendCh<Msg>, engine: Arc<DB>, _: &str) {
@@ -64,12 +86,21 @@ mod imp {
                     if let Some(v) = engine.get_statistics() {
                         info!("{}", v)
                     }
-                    profiling::print_prof();
+                    print_prof();
                 }
                 SIGUSR2 => profiling::dump_prof(None),
                 // TODO: handle more signal
                 _ => unreachable!(),
             }
+        }
+    }
+
+    #[cfg(test)]
+    mod tests {
+        #[test]
+        fn test_stats_print() {
+            // just print the data, ensure it doesn't core.
+            super::print_prof()
         }
     }
 }
